+++ packages/database/src/__tests__/modelClasses/EntityParentChildRelation/setup.js
@@ -1,0 +1,28 @@
+import { ENTITIES, ENTITY_RELATIONS, PROJECTS } from './fixtures';
+import { buildAndInsertProjectsAndHierarchies, clearTestData } from '../../../server/testUtilities';
+import { EntityHierarchySubtreeRebuilder } from '../../../server/changeHandlers/entityHierarchyCacher/EntityHierarchySubtreeRebuilder';
+
+export const setupTestData = async models => {
+  await clearTestData(models.database);
+
+  const projectsForInserting = PROJECTS.map(project => {
+    const relationsInProject = ENTITY_RELATIONS.filter(
+      relation => relation.hierarchy === project.code,
+    );
+    const entityCodesInProject = relationsInProject.map(relation => relation.child);
+    const entitiesInProject = entityCodesInProject.map(entityCode =>
+      ENTITIES.find(entity => entity.code === entityCode),
+    );
+    return { ...project, entities: entitiesInProject, relations: relationsInProject };
+  });
+
+  await buildAndInsertProjectsAndHierarchies(models, projectsForInserting);
+  const projects = await Promise.all(
+    projectsForInserting.map(project => models.project.findOne({ code: project.code })),
+  );
+
+  const hierarchyCacher = new EntityHierarchySubtreeRebuilder(models);
+  for (const project of projects) {
+    await hierarchyCacher.buildAndCacheProject(project);
+  }
+};
